[TASK] Remove legacy DataMapper
A DataMapper is used in the context of Extbase
which does a configurable mapping between
a property of a class and a field in the
storage (e.g. database).

In the first implementation, the code was widely taken
taken and adapted from Extbase to stay in the
same spirit API wise.
However, the concept of mapping does not apply
to Vidi, since it is implicitly done upon the
field name.

Change-Id: Ie2fed6f86b10b538a2cf2e09ab146e23156549de
Releases: 0.4.0
Resolves: #59078