[BUGFIX] Respect data type in FacetSuggestionService::getSuggestions()
Until now an exception was thrown if a Facet is configured for a field of a
datatype (table name) other than ModuleLoader::getDataType.

This only happens in FacetSuggestionService::getSuggestions() if [1] the field
is not configured as text area and [2] the number of suggestions is lower or
equal the configured limit.

Resolves #7